# Implement inheritance-based export system