--- a/src/config.rs
+++ b/src/config.rs
@@ -3,4 +3,4 @@ pub mod item_filter;
 
 // Re-export key types for convenience
 pub use global::{GlobalConfig, ScanConfig, VisualizationConfig};
-pub use item_filter::{ItemFilterConfig, ItemTypeConfig, ExclusionRules, ValidationRules};
+pub use item_filter::{ItemFilterConfig, ItemTypeConfig, ExclusionRules};